feature_extraction/requirements.txt
google-cloud-storage
pandas
numpy